fix: make template for register and login
